Refactor: Deconstruct Landing, Messages, and Sessions God Components
Resolves #605

- Extracted Landing sections into Hero, Features, Communities, Testimonials.

- Abstracted Supabase/real-time logic into useMessages.ts and separated UI.

- Abstracted sessions logic into useSessions.ts and separated UI.

- Cleaned up syntax error in Room.tsx.
